chimera-core/requirements.txt
# Chimera Core - Python 3.12 Worker Dependencies
# Complete production requirements for stealth browser automation worker

# ============================================================================
# gRPC & Protocol Buffers (Brain Communication)
# ============================================================================
grpcio>=1.60.0,<2.0.0
grpcio-tools>=1.60.0,<2.0.0
protobuf>=4.25.0,<5.0.0

# ============================================================================
# Browser Automation (Stealth Worker Swarm)
# ============================================================================
playwright>=1.40.0,<2.0.0

# ============================================================================
# Image Processing (Screenshots for Vision Service)
# ============================================================================
Pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0

# ============================================================================
# HTTP Clients & Stealth
# ============================================================================
requests>=2.31.0,<3.0.0
httpx>=0.27.0,<1.0.0
curl_cffi>=0.5.10,<1.0.0  # Chrome TLS fingerprint stealth

# ============================================================================
# Queue & State Management
# ============================================================================
redis>=5.0.1,<6.0.0  # For worker queue coordination

# ============================================================================
# Database Persistence (PostgreSQL)
# ============================================================================
psycopg2-binary==2.9.9  # PostgreSQL adapter for mission results

# ============================================================================
# Logging & Monitoring
# ============================================================================
loguru>=0.7.2,<1.0.0

# ============================================================================
# Async & Utilities
# ============================================================================
aiofiles>=23.2.1,<24.0.0
python-dotenv>=1.0.0,<2.0.0

# ============================================================================
# Stealth & Randomization (Human-like Behavior)
# ============================================================================
# Note: Python's built-in random and math modules are sufficient
# For advanced noise/diffusion paths, consider: noise>=1.2.2 (optional)
